KD-2996 - Delete component parts when biblio is deleted
When merging biblios with the deduplicator, we cannot just
delete component parts, as the deduped biblios may have exact
same 001+003 fields, which is what component records use to
know what biblio they belong to.

TODO: Check the 001 and 003 fields, and if they differ between
the biblios to be deleted vs. the biblio to be kept, then maybe delete.