KD-3816 (squashable) Add support for buyer assigned id
Vendor EDI accounts allows mapping the booksellers based on buyer
assigned ID instead of vendor assigned ID in shipment notices,
so we'll also support that. Buyer assigned IDs are less likely
to change when baskets change on the booksellers end, so there
will be no need to change the mappings in Koha in such cases.