KD-856 Leave items "in transit" on self-checkin with SIP
The items checked in by the patrons with self check-in machines (SIP)
will always be "in transit" until the staff has checked them in again with
Staff Client. The idea is to delay sending the hold pickup notices until the
library staff has had time to shelf the holds for pickup. Sending notices
immediately on self check-in is especially problematic during self service
hours in the libraries.

The feature is controlled by "RequireSCCheckInBeforeNotifyingPickups"
system preference and it defaults to normal Koha behaviour (send notices
immediately). You will need to set the system preference in order for the
notices to be delayed.