KD-3945 Do not change dateaccessioned when receiving items
Handling dates in finnishreceive.pl is not logical. Dateaccessioned
(along with replacementpricedate and datelastseen) gets the date set by the
user on receive-page, while datereceived gets current date regardless of the
user's selection. Dateaccessioned should not change at all when receiving
items, since it's generally just the date that an order was placed on item.

We will change receiving to not tamper with dateaccesioned, and instead set
datereceived (along with replacementpricedate and datelastseen) according
to the user's selection on the receive-page.