KD-3947 - Language codes as authorised values
When a (sub)field is set so authorised values are used for it, the
cataloguers pick the values from a pull-down in the editor.
The MARCXML gets the authorised value, but when viewing the record details,
the authorised value is replaced by the description before the record is
passed on to the XSLT templates. (See transformMARCXML4XSLT)

This changes the language XSLT to only look at the first 3 characters
of the value, so if the description's first characters match the
language code, it'll work, and if the 041 subfields aren't set up
to take their values from authorised values, it'll also work.

You'll have to create the language codes so the value is the code, and
the description (in both intranet and opac) is prefixed with the 3-character
language code, eg.

code   description    description (OPAC)
eng    eng - English  eng - English
tlh    tlh - Klingon  tlh - Klingon