Bug 24683: Fix for smart rules applied when forming "AnyAvailable" state
When there was a call to ItemsAnyAvailableForHold, inside of this sub
there was no effect from main smart rules set: i.e. call to
"CanItemBeReserved" was absent totally.

Because of this, there was a problem: none of the items was allowed to
hold, when there are few different item types for one biblio, and, let's
say, those which allowed for a hold on load, and another not allowed to
hold by smart rule (for example 0 "Holds per record"), but not on loan.
Such set made for example "all unavailable" and on shelf holds should
work, but it failed. Note: more detailed explanation down below in "how
to reproduce" description.

Call to sub "CanItemBeReserved" added and check for "...->{status} ne 'OK'"
now has an impact on the result.

How to reproduce:

1. Add 2 smart rules (/cgi-bin/koha/admin/smart-rules.pl) with "on shelf
   holds": "if all unavailable" for all rules, and set "holds per record"
   to 2 for "books" and "0" for "computer files".

2. Create only 2 items for one biblio, but different types, "book"
   and "computer file". For example in misc4dev env:
   /cgi-bin/koha/cataloguing/additem.pl?biblionumber=1#additema

3. Check out that item of type "book" to some person, for example,
   in misc4dev:
   /cgi-bin/koha/circ/circulation.pl?borrowernumber=2&barcode=3999900000001

4. Open reserve/request, for example, for item 1 and patron 1 in misc4dev
   env (/cgi-bin/koha/reserve/request.pl?biblionumber=1&borrowernumber=1)

5. It does not allow to hold, both red crossed, but computer file says
   "Exceeded max holds per record" because of "0" limit set on step 1.

6. Apply the patch.

7. Reload page on step 5 and see that "book" will be available for hold,
   but "computer file" still will be red-crossed "Exceeded max holds
   per record", now that's correct because both items unavailable:
   one because on load, another because of "0" limit for computer files.

8. Check-in book from step 3 so it will be returned to the library,

9. Reload page on step 5 and see that again no any holds available,
   but it's now also correct: "book" now returned but "on shelf holds"
   set to "if all unavailable".